Revises comments within NIKO_WUZ_HERE_LOL.py and Virus.py so they are a consistent style and less wordy.
Hopefully this new format will be friendly to beginners! I standardized the docstrings that describe each method. I also included a 'How to use:' in each docstring so beginners do not have to guess how these methods are used. I also just made a few tweaks to Zaynab's language here and there so it would be extremely beginner-friendly, basically giving them instructions on how to succeed, but also noting for students who are interested 'There are many ways to solve this puzzle!' in Virus.py.